+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 * Add validation error code `Transaction.ProductSkuIsInvalid`
 * Add 'RiskThreshold' gateway rejection reason
+* Add `processed_with_network_token` to `Transaction`
+* Add `is_network_tokenized` to `CreditCard`
 
 ## 4.2.0
 * Add `retrieval_reference_number` to `Transaction`